fixed printf format warning on 64 bit platforms
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@38999 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/generic/htmllbox.cpp
@@ -301,7 +301,7 @@ void wxHtmlListBox::CacheItem(size_t n) const
 
         // set the cell's ID to item's index so that CellCoordsToPhysical()
         // can quickly find the item:
-        cell->SetId(wxString::Format(_T("%u"), n));
+        cell->SetId(wxString::Format(_T("%lu"), (unsigned long)n));
 
         cell->Layout(GetClientSize().x - 2*GetMargins().x);
